accounting/requirements.txt
asgiref==3.6.0
Django==3.1.7
django-cors-headers==3.5.0
django-mysql==3.9.0
mysqlclient==2.0.1
pytz==2023.3
sqlparse==0.4.4
djangorestframework==3.12.4